reorganize files, start progress bar
