+++ src/components/Atoms/ProgressBar.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function ProgressBar() {
+  return (
+    <div>ProgressBar</div>
+  );
+}
